feat: rename blockstore package
