+++ venus-shared/blockstore/autobatch_test.go
@@ -1,4 +1,4 @@
-package blockstoreutil
+package blockstore
 
 import (
 	"context"
