+++ venus-shared/blockstore/badger.go
@@ -1,4 +1,4 @@
-package blockstoreutil
+package blockstore
 
 import (
 	"context"
