+++ venus-shared/blockstore/safe_cid_set.go
@@ -1,4 +1,4 @@
-package blockstoreutil
+package blockstore
 
 import (
 	"sync"
